+++ app/workers/delay_henka/apply_actions_worker.rb
@@ -3,9 +3,10 @@ module DelayHenka
 
     include Sidekiq::Worker
 
-    def perform
+    def perform(time_zone)
       ScheduledAction.staged
         .where('schedule_at <= ?', Time.current)
+        .where(time_zone: time_zone)
         .includes(:actionable)
         .find_each(&:apply_action)
     end