+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## Version 0.6.2
+* require time_zone as parameter on schedule method for ScheduledChange and ScheduledAction.
+* Add time_zone to query inside ApplyActionsWorker and ApplyChangesWorker.
+* wrap ApplyActionsWorker and ApplyChangesWorker in UpdatesOnValidTimeZonesWorker.
+
 ## Version 0.6.1
 * Introduce time_zone as parameter on schedule method for ScheduledChange and ScheduledAction.